refactor(quiz): update import paths for QuizVariant and create a new constants file for better organization
