+++ features/answer-quiz/constants/quizVariant.ts
@@ -2,4 +2,4 @@ export enum QuizVariant {
   UNANSWERED = 'unanswered',
   CORRECT = 'correct',
   INCORRECT = 'incorrect',
-} 
+}
